Ad warnings, be gone.

+++ src/tools/ad/System.cpp
@@ -235,7 +235,7 @@ uint32 ReadMapDBC()
         map_ids[x].id = dbc.getRecord(x).getUInt(0);
         strcpy(map_ids[x].name, dbc.getRecord(x).getString(1));
     }
-    printf("Done! (%u maps loaded)\n", map_count);
+    printf("Done! (%u maps loaded)\n", (unsigned int)map_count);
     return map_count;
 }
 
@@ -260,7 +260,7 @@ void ReadAreaTableDBC()
 
     maxAreaId = dbc.getMaxId();
 
-    printf("Done! (%u areas loaded)\n", area_count);
+    printf("Done! (%u areas loaded)\n", (unsigned int)area_count);
 }
 
 void ReadLiquidTypeTableDBC()
@@ -281,7 +281,7 @@ void ReadLiquidTypeTableDBC()
     for(uint32 x = 0; x < LiqType_count; ++x)
         LiqType[dbc.getRecord(x).getUInt(0)] = dbc.getRecord(x).getUInt(3);
 
-    printf("Done! (%u LiqTypes loaded)\n", LiqType_count);
+    printf("Done! (%u LiqTypes loaded)\n", (unsigned int)LiqType_count);
 }
 
 //
